Add counter-lazy function
This function prevents rat from stay at the lowest position of the
escalator, where always has lowest speed of the motor(rat is also
a smart creature...). We increase the speed level of the escalator
every 10 seconds, unless the speed reach the limit we will keep
increase it, and request for user-specified speed every 30 seconds
, note that the speed sent from PC is evaluated, which means that
we analyze which position the rat stay longest.